// src/app/core/enumerations/message-action.enum.ts
export enum MessageActionEnum {
    Page = 'page',
    Confirm = 'confirm',
    Message = 'message',
    Report = 'report',
    Upload = 'upload',

    Disconnect = 'disconnect',
    ClosePage = 'closepage',
    RenderPage = 'renderpage',
    RefreshPage = 'refreshpage',
    ChangeCtlProp = 'changectlprop',
    DeleteRows = 'deleterows',
    OpenFile = 'openfile',
    SaveFile = 'savefile',
    LoadingStart = 'loading_start',
    LoadingStop = 'loading_stop',
    OpenLink = 'openlink',

    UpdateNotifications = 'updatenotifications'
}